requirements.txt
numpy
pandas
lxml
scrapy_playwright
scrapy
playwright
termcolor
tqdm
opencv-python
pillow